Added participants table with required information
Participants are separate from users so it will be possible for one person to register multiple people for the con.
Implementation should follow the pattern "Participants can be Users but don't have to be".
Events have many participants, participants can belong to many events.
Users can have many participants, participants belong to one user.
For future-proofing, it should be possible to change ownership of a participant to a different user, e.g. if I register
multiple people this year, but can't come next year so group registration is handeled by someone else, we don't have to add the same participants information again.
Also useful if somebody wants to register themselves at some point later on instead of being registered as part of a group.

Link between User and Participant should happen via e-mail as this is the only thing present in the users and participants table.