Added a form to event registrations which contains one row per participant to be added. Made the form respond to an ajax call from a link to add an additional participant row. This way, the form can be dynamically extended to how many participants are needed for a given registration.
TODO: The backend side of it all is not yet done.
We still need to handle the data coming from the form at controller and model level.
Also, it might be nice to be able to remove rows again if clicked on the button too often.
Styling the whole thing will also have to wait for now...