Added submit button to preview. Problem: All the data submitted is passed as Query string instead of in hidden forms as the api documentation of button_to claims. Very ugly and unsafe in my opinion.
Design Flaw just uncovered: we're missing a join table to join participants to the event registration. using event registration as join table to link participants to events will not work when
we try to group multiple participants in one registration e.g. for billing and stuff later...

Back to database design.