Removed email field from participants. Usually, each participant tends to register for himself. In this case, the Participant also has to register or log in as a User which already has an email address.
In case of a group registration, the User doing the registration is responsible for his group.
Therefore, all the mails should go to this responsible user.
In most cases of group registrations, it will be parents registering their family. Most kids won't even have an email.

This should be sufficient for now.
It could be desirable in the future to allow sending of mails to all participants and not to the user - optimization needed here.