Added a form to enter e-mail. The controller method send_registration_link right now creates an empty registration in the database with only the registration_token set. No Mailer is configured yet.
We encounter problems in our admin views. When trying to display all registrations for an event, the call to event_registration.user.email returns nil as there is no user associated with this new registration yet.
A bit more work is needed here to have the view either ignore those pending late registrations or display some message instead of breaking on nil.
After that, we can configure the mailer.

Some progress on #12