default_url_options defined on mailer
Apparently, the config.action_mailer.default_url_options are not read from the environments.
ApiDoc for rails tells me we have to define this in application.rb - which does not work
if you have different hosts in different environments.

So the solution seems to be defining a default_url_options method on the mailer which reads the current Rails.application.config.action_mailer.default_url_options